Add new app helper and rename file
